Use node-search-engine from ai-workflow-builder to search nodes, consolidate search-nodes-tools into node-execution-tools
