+++ packages/@n8n/ai-workflow-builder.ee/src/index.ts
@@ -12,6 +12,9 @@ export type {
 	StreamWriter,
 } from './assistant';
 
+export { NodeSearchEngine } from './tools/engines/node-search-engine';
+export type { NodeSearchResult } from './types/nodes';
+
 // Code builder utilities for MCP integration
 export {
 	NodeTypeParser,
